% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/assess.R
\name{assess}
\alias{assess}
\title{Assess Generalizability of Randomized Sample to Population}
\usage{
assess(
  data,
  guided = TRUE,
  sample_indicator,
  covariates,
  estimation_method = "lr",
  disjoint_data = TRUE,
  trim_pop = FALSE
)
}
\arguments{
\item{data}{data frame comprised of "stacked" sample and target population data}

\item{guided}{logical. Default is TRUE. If FALSE, then user must enter all arguments in function to bypass guided mode}

\item{sample_indicator}{variable name denoting sample membership (1 = in sample, 0 = out of sample)}

\item{covariates}{vector of covariate names in data set that predict sample membership}

\item{estimation_method}{method to estimate the probability of sample membership (propensity scores). Default is logistic regression ("lr"). Other methods supported are Random Forests ("rf") and Lasso ("lasso")}

\item{disjoint_data}{logical. If TRUE, then sample and population data are considered disjoint. This affects calculation of the weights - see details for more information.}

\item{trim_pop}{logical. If TRUE, then population data are subset to exclude individuals with covariates outside bounds of sample covariates.}
}
\value{
returns generalizeR_assess object that includes the generalizability index, propensity scores, and a covariate table
}
\description{
This function, given a stacked data frame containing both sample and population data, assesses the generalizability of the sample to the population on given covariates.
}
