% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/assess_wrap.R
\name{assess_wrap}
\alias{assess_wrap}
\title{Assess the Generalizability of a Sample to a Population}
\usage{
assess_wrap(sample, population, join_var = NULL, grouping_var = NULL)
}
\arguments{
\item{sample}{a data frame with a column of IDs}

\item{population}{a data frame with a column of IDs, columns for any stratifying variables of interest, and (if applicable) a grouping variable}

\item{join_var}{defaults to NULL; name of the ID variable in double quotes}

\item{grouping_var}{defaults to NULL; name of a grouping variable in double quotes (not required)}
}
\description{
This function is a wrapper for the function 'assess()' which may make assessment more user-friendly and flexible.
}
\details{
Given a data frame or tibble consisting of a column of some ID variable (the sample) and a data frame or tibble of the same ID variable and stratifying variables of interest, this function will assess the generalizability of the sample to the population.

The user can also specify a grouping variable to assess the generalizability of their sample across population groups -- for example, across US states. This will produce multiple generalizability index values, one per level of the grouping variable.
}
