% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/assess.R
\name{.assess.guided}
\alias{.assess.guided}
\title{Guided Version of \code{Assess()}, Allowing Guiding User Input to Assess Generalizability of Randomized Sample to Population}
\usage{
.assess.guided(data)
}
\arguments{
\item{data}{data.frame, a data frame containing both sample and population data}
}
\value{
A five-element list containing the name of an indicator variable for whether
a row is in the sample or the (disjoint) population,
the corresponding data.frame covariate columns of interest,
a boolean indicating whether the sample and population data are disjoint,
a boolean indicating whether the population data should be trimmed to reflect
bounds of covariates within the sample, and a string identifying the chosen estimation method
}
\description{
Intended only to be called within \code{assess()}, not as a standalone function
}
\keyword{internal}
