% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stratify.R
\name{.check.factor.levels}
\alias{.check.factor.levels}
\title{Internal function that checks whether any column in a dataset is a factor variable with more than a given number of levels}
\usage{
.check.factor.levels(data, maxlevels = 4L)
}
\arguments{
\item{data}{data.frame, a dataframe of factor variables of interest}

\item{maxlevels}{integer, a maximum permissible number of levels}
}
\value{
invalid_factors, list of variable names with more than the permitted number of levels
}
\description{
Intended only to be called within \code{stratify}, not as a standalone function
}
\keyword{internal}
