% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stratify.R
\name{.check.no.na.cols}
\alias{.check.no.na.cols}
\title{Internal function that checks whether a list provided is empty (avoiding conflict w/ check on >4 level factor categorical vars)}
\usage{
.check.no.na.cols(na_col_list)
}
\arguments{
\item{na_col_list}{list, a list of char corresponding to dataframe columns where all rows are NA}
}
\value{
logical, whether the inputted list is empty
}
\description{
Intended only to be called within \code{stratify}, not as a standalone function
}
\keyword{internal}
