% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{.generate.ps}
\alias{.generate.ps}
\title{Generate Sample Participation Probabilities}
\usage{
.generate.ps(data, sample_indicator, covariates, estimation_method)
}
\arguments{
\item{data}{data frame comprised of "stacked" sample and target population data}

\item{sample_indicator}{variable name denoting sample membership (1 = in sample, 0 = out of sample)}

\item{covariates}{vector of covariate names in data set that predict sample membership}

\item{estimation_method}{method to estimate the probability of sample membership. Default is logistic regression ("lr").Other methods supported are Random Forests ("rf") and Lasso ("lasso")}
}
\value{
sample participation probabilities for each unit in the data frame
}
\description{
This function is designed for use within 'weighting()' and 'assess()'.
}
