% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/recruit.R
\name{.get.sample.size}
\alias{.get.sample.size}
\title{Helper function that asks user to specify sample size for guided version}
\usage{
.get.sample.size(valid_inputs)
}
\description{
Helper function that asks user to specify sample size for guided version
}
