% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stratify.R
\name{.make.cont.data.tbl}
\alias{.make.cont.data.tbl}
\title{Internal function that creates and formats a summary table of the continuous variables within the provided dataset}
\usage{
.make.cont.data.tbl(cont_data)
}
\arguments{
\item{cont_data}{data.frame, a dataframe of continous variables}
}
\value{
The function returns a dataframe with summary statistics of the inputted data
}
\description{
Intended only to be called within \code{stratify}, not as a standalone function
}
\keyword{internal}
