% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{.make.covariate.table}
\alias{.make.covariate.table}
\title{Create Covariate Balance Table}
\usage{
.make.covariate.table(
  data,
  sample_indicator,
  covariates,
  sample_weights = NULL,
  estimation_method = "lr",
  disjoint_data = TRUE
)
}
\arguments{
\item{data}{Dataframe comprised of "stacked" sample and target population data}

\item{sample_indicator}{Binary variable denoting sample membership (1 = in sample, 0 = out of sample)}

\item{covariates}{Vector of covariates in dataframe that predict sample membership}

\item{sample_weights}{Name of column in dataframe holding weights for calculating weighted sample means of covariates in dataframe. If NULL, sample means are unweighted.}

\item{estimation_method}{Method to estimate the probability of sample membership. Default is logistic regression ("lr"). Other methods supported are Random Forests ("rf") and Lasso ("lasso").}

\item{disjoint_data}{Logical. Defaults to TRUE. If TRUE, then sample and population data are considered disjoint. This affects calculation of the weights.}
}
\description{
This function is designed for use within \code{weighting()} and \code{assess()}.'
}
