% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/recruit.R
\name{.make.recruitment.lists}
\alias{.make.recruitment.lists}
\title{Internal function that performs recruitment calculations}
\usage{
.make.recruitment.lists(stratify_output)
}
\description{
Internal function that performs recruitment calculations
}
