% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/recruit.R
\name{.make.recruitment.table}
\alias{.make.recruitment.table}
\title{Internal function that formats a summary table of stata-based recruitment results for \code{recruit()}}
\usage{
.make.recruitment.table(stratify_output, sample_size)
}
\arguments{
\item{stratify_output}{list of class generalizer_stratify}

\item{sample_size}{integer, total sample size selected to guide recruitment}
}
\value{
The function returns a list of formatted recruitment results for user interpretation: a table of results, a corresponding kable, and a the number of units selected from each stratum
}
\description{
Intended only to be called within \code{recruit()}, not as a standalone function
}
\keyword{internal}
