% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stratify.R
\name{.make.var.overview}
\alias{.make.var.overview}
\title{Internal function that creates and formats a table of the variables in the dataset provided}
\usage{
.make.var.overview(dataset, print_to_console = FALSE)
}
\arguments{
\item{dataset}{data.frame, a dataframe of continous variables}

\item{print_to_console}{logical, whether the variable summary table should be printed to the console, defaults to false}
}
\value{
None
}
\description{
Intended only to be called within \code{stratify}, not as a standalone function
}
\keyword{internal}
