% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stratify.R
\name{.n.strata.less.than.max}
\alias{.n.strata.less.than.max}
\title{Internal function that ensures that the number of strata requested does not exceed the maximum possible (maximum n-2 where n is number of observations w/ no missing variables of interest)}
\usage{
.n.strata.less.than.max(n_strata, data_interest, variables, idvar)
}
\arguments{
\item{n_strata}{integer, a number of strata in which to divide to cluster population}

\item{variables}{character, provide a character vector of the names of stratifying variables (from population data frame)}

\item{idvar}{character, provide a character vector of the name of the ID variable (from population data frame)}

\item{data}{data.frame, a data.frame object with the population of interest (rows are observations)}
}
\value{
The function returns a boolean value indicating whether the number of observations with non-missing variables of interests exceeds 1 plus the number of strata requested
}
\description{
Intended only to be called within \code{stratify}, not as a standalone function
}
\keyword{internal}
