% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/recruit.R
\name{.round.preserve.sum}
\alias{.round.preserve.sum}
\title{Internal function used in the calculations for creating recruitment lists, under \code{.make.recruitment.lists()}. Rounds number of observations to be recruited from each stratum to a whole number, while adjusting rounded results to preserve the total.}
\usage{
.round.preserve.sum(x, digits = 0)
}
\arguments{
\item{x}{list of numeric values; enumerates the possible number of units to recruit, based on the population size}

\item{digits}{integer; signals number of 0 digits to place after each numeric in x (e.g. digits = 2 will multiply x by 100)}
}
\value{
The function returns a list of integers, which will represent the number of observations to be sampled from each stratum
}
\description{
Intended only to be called within \code{recruit()}, not as a standalone function
}
\keyword{internal}
