% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/assess.R
\name{.select.covariates}
\alias{.select.covariates}
\title{Part of Guided Version of \code{Assess()}, guiding users through the selection of covariates of interest}
\usage{
.select.covariates(data)
}
\arguments{
\item{data}{data.frame, a data frame containing both sample and population data}
}
\value{
data.frame column with an indicator variable designating whether an observation
is part of their sample or the (disjoint) population
}
\description{
Intended only to be called within \code{assess()}, not as a standalone function
}
\keyword{internal}
