% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stratify.R
\name{.select.list.stratify}
\alias{.select.list.stratify}
\title{Internal function that allows users to choose the variables of interest for stratifying observations in their dataset}
\usage{
.select.list.stratify(data)
}
\arguments{
\item{data}{data.frame, a dataframe of covariates of potential interest}
}
\value{
data.frame, a dataframe of user-selected covariates of interest
}
\description{
Intended only to be called within the guided version of \code{stratify}, not as a standalone function
}
\keyword{internal}
