% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/assess.R
\name{.select.method}
\alias{.select.method}
\title{Part of Guided Version of \code{Assess()}, guiding users through the selection of
a generalizability assessment method (logistic regression, random forest,
or lasso regression)}
\usage{
.select.method()
}
\value{
string referring to the chosen assessment method
}
\description{
Intended only to be called within \code{assess()}, not as a standalone function
}
\keyword{internal}
