% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stratify.R
\name{.stratify.calculate}
\alias{.stratify.calculate}
\title{Internal function that performs the actual stratification calculations for \code{stratify()}}
\usage{
.stratify.calculate(data, n_strata, variables, idvar, verbose)
}
\description{
Intended only to be called within \code{stratify_guided} and \code{stratify_unguided}, not as a standalone function
}
\keyword{internal}
