% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/assess.R
\name{.trim.pop}
\alias{.trim.pop}
\title{Subset Population so Population Covariates are within bounds of Sample Covariates}
\usage{
.trim.pop(data, sample_var, covariates)
}
\arguments{
\item{data}{data frame comprised of "stacked" sample and target population data}

\item{sample_var}{variable name denoting sample membership (1 = in sample, 0 = out of sample)}

\item{covariates}{vector of covariate names in data set that predict sample membership}
}
\value{
\code{trim_pop} returns a data frame, where the target population covariates do not exceed the bounds of the sample covariates
}
\description{
Subset Population so Population Covariates are within bounds of Sample Covariates
}
