% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/assess.R
\name{.yes.no}
\alias{.yes.no}
\title{Part of Guided Version of \code{Assess()}, helper function that aids in ensuring that
users have entered correct input in response to a Yes/No question}
\usage{
.yes.no(message = "")
}
\arguments{
\item{message}{string, a Yes/No question to be printed to the user's console}
}
\value{
boolean referring to the user's valid response
}
\description{
Intended only to be called within \code{assess()}, not as a standalone function
}
\keyword{internal}
