% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/assess.R, R/generate_ps.R
\name{generate_ps}
\alias{generate_ps}
\title{Generate Sample Participation Probabilities}
\usage{
generate_ps(data, trial, selection_covariates, selection_method)

generate_ps(data, trial, selection_covariates, selection_method)
}
\arguments{
\item{data}{data frame comprised of "stacked" trial and target population data}

\item{trial}{variable name denoting binary trial participation (1 = trial participant, 0 = not trial participant)}

\item{selection_covariates}{vector of covariate names in data set that predict trial participation}

\item{selection_method}{method to estimate the probability of trial participation. Default is logistic regression ("lr").Other methods supported are Random Forests ("rf") and Lasso ("lasso")}
}
\value{
sample participation probabilities for each unit in the data frame

sample participation probabilities for each unit in the data frame
}
\description{
This function is designed for use within 'weighting()' and 'assess()'.

This function is designed for use within 'weighting()' and 'assess()'.
}
