% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generalize.R
\name{print.generalize}
\alias{print.generalize}
\title{Formats and prints results from a \code{generalize} object generated by the \code{generalize()} function}
\usage{
\method{print}{generalize}(x, ...)
}
\arguments{
\item{x}{generalize object, an object generated by the \code{generalize()} function}
}
\description{
Formats and prints results from a \code{generalize} object generated by the \code{generalize()} function
}
