% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/assess.R
\name{print.generalize_assess}
\alias{print.generalize_assess}
\title{Formats and prints results from a \code{generalizer_assess} object generated by the \code{assess()} function}
\usage{
\method{print}{generalize_assess}(x, ...)
}
\arguments{
\item{x}{generalizer_assess object, an object generated by the \code{assess()} function}
}
\description{
Formats and prints results from a \code{generalizer_assess} object generated by the \code{assess()} function
}
