% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/assess.R
\name{print.generalizeR_assess}
\alias{print.generalizeR_assess}
\title{Print method for "generalizeR_assess" class}
\usage{
\method{print}{generalizeR_assess}(x, ...)
}
\arguments{
\item{x}{An object of class "generalizeR_assess"}

\item{...}{Other arguments passed to or from other methods}
}
\value{
The generalizability index for the sample along with details about the sample
}
\description{
Print method for "generalizeR_assess" class
}
