% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/recruit.R
\name{print.generalizeR_recruit}
\alias{print.generalizeR_recruit}
\title{Print method for "generalizeR_recruit" class}
\usage{
\method{print}{generalizeR_recruit}(x, ...)
}
\arguments{
\item{x}{An object of class "generalizeR_recruit"}

\item{...}{Other arguments passed to or from other methods}
}
\description{
Print method for "generalizeR_recruit" class
}
