% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stratify.R
\name{print.generalizeR_stratify}
\alias{print.generalizeR_stratify}
\title{Print method for "generalizeR_stratify" class}
\usage{
\method{print}{generalizeR_stratify}(x, ...)
}
\arguments{
\item{x}{An object of class "generalizeR_stratify"}

\item{...}{Other arguments passed to or from other methods}
}
\value{
A list of the covariates used to stratify, the number of strata, and the number of observations dropped due to missing data
}
\description{
Print method for "generalizeR_stratify" class
}
