% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/weighting.R
\name{print.generalizeR_weighting}
\alias{print.generalizeR_weighting}
\title{Print method for "generalizeR_weighting" class}
\usage{
\method{print}{generalizeR_weighting}(x, ...)
}
\arguments{
\item{x}{An object of class "generalizeR_weighting"}

\item{...}{Other arguments passed to or from other methods}
}
\value{
A summary of the weighted dataset
}
\description{
Print method for "generalizeR_weighting" class
}
