% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generalize.R
\name{print.summary.generalize}
\alias{print.summary.generalize}
\title{Formats and prints results from a \code{summary.generalize} object generated by the \code{summary.generalize()} function}
\usage{
\method{print}{summary.generalize}(x, ...)
}
\arguments{
\item{x}{summary.generalize object, an object generated by the \code{summary.generalize()} function}
}
\description{
Formats and prints results from a \code{summary.generalize} object generated by the \code{summary.generalize()} function
}
