% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/assess.R
\name{print.summary.generalize_assess}
\alias{print.summary.generalize_assess}
\title{Formats and prints results from a \code{summary.generalize_assess} object generated by the \code{summary.generalize_assess()} function}
\usage{
\method{print}{summary.generalize_assess}(x, ...)
}
\arguments{
\item{x}{summary.generalize_assess object, an object generated by the \code{summary.generalize_assess()} function}
}
\description{
Formats and prints results from a \code{summary.generalize_assess} object generated by the \code{summary.generalize_assess()} function
}
