% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/weighting.R
\name{print.summary.generalizeR_weighting}
\alias{print.summary.generalizeR_weighting}
\title{Print method for "summary.generalizeR_weighting" class}
\usage{
\method{print}{summary.generalizeR_weighting}(x, ...)
}
\arguments{
\item{x}{An object of class "summary.generalizeR_weighting"}

\item{...}{Other arguments passed to or from other methods}
}
\value{
A summary of the weighted and unweighted dataset for propensity scores, covariates and (optionally) TATE
}
\description{
Print method for "summary.generalizeR_weighting" class
}
