% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stratify.R
\name{stratify}
\alias{stratify}
\title{Stratify a Population Data Frame}
\usage{
stratify(data, guided = TRUE, n_strata = NULL, variables = NULL, idnum = NULL)
}
\arguments{
\item{data}{The R object containing your population data frame}

\item{guided}{logical; defaults to TRUE. Whether the function should be guided (ask questions and behave interactively throughout) or not. If set to FALSE, must provide values for other arguments below}

\item{n_strata}{defaults to NULL. If guided is set to FALSE, must provide a number of strata to cluster population into}

\item{variables}{defaults to NULL. If guided is set to FALSE, must provide a character vector of the names of stratifying variables (from population data frame)}

\item{idnum}{defaults to NULL. If guided is set to FALSE, must provide a character vector of the name of the ID variable (from population data frame)}
}
\value{
A list of class "generalizer_output" that can be given to other functions in the package
}
\description{
This function takes as input any data frame that you want to stratify into clusters. Typically, the goal of such stratification is sampling for generalizability. This function, and the others in this package, are designed to mimic the website https://www.thegeneralizer.org/.
}
