% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generalize.R
\name{summary.generalize}
\alias{summary.generalize}
\title{Summarizes results from a \code{generalizer_assess} object generated by the \code{generalize()} function}
\usage{
\method{summary}{generalize}(object, ...)
}
\arguments{
\item{object}{generalize object, an object generated by the \code{generalize()} function}
}
\value{
an object of class \code{summary.generalize}
}
\description{
Summarizes results from a \code{generalizer_assess} object generated by the \code{generalize()} function
}
