% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/assess.R
\name{summary.generalize_assess}
\alias{summary.generalize_assess}
\title{Summarizes results from a \code{generalizer_assess} object generated by the \code{assess()} function}
\usage{
\method{summary}{generalize_assess}(object, ...)
}
\arguments{
\item{object}{generalizer_assess object, an object generated by the \code{assess()} function}
}
\value{
an object of class \code{summary.generalize_assess}
}
\description{
Summarizes results from a \code{generalizer_assess} object generated by the \code{assess()} function
}
