% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/assess.R
\name{summary.generalizeR_assess}
\alias{summary.generalizeR_assess}
\title{Summary method for "generalizeR_assess" class}
\usage{
\method{summary}{generalizeR_assess}(x, ...)
}
\arguments{
\item{x}{An object of class "generalizeR_assess"}

\item{...}{Other arguments passed to or from other methods}
}
\description{
Summary method for "generalizeR_assess" class
}
