% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/weighting.R
\name{summary.generalizeR_weighting}
\alias{summary.generalizeR_weighting}
\title{Summary method for "generalizeR_weighting" class}
\usage{
\method{summary}{generalizeR_weighting}(x, ...)
}
\arguments{
\item{x}{An object of class "generalizeR_weighting"}

\item{...}{Other arguments passed to or from other methods}
}
\description{
Summary method for "generalizeR_weighting" class
}
