% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/assess.R
\name{assess}
\alias{assess}
\title{Assess Generalizability of Randomized Trial to Population}
\usage{
assess(
  trial,
  selection_covariates,
  data,
  selection_method = "lr",
  is_data_disjoint = TRUE,
  trim_pop = FALSE,
  seed = 12222
)
}
\arguments{
\item{trial}{variable name denoting binary trial participation (1 = trial participant, 0 = not trial participant)}

\item{selection_covariates}{vector of covariate names in data set that predict trial participation}

\item{data}{data frame comprised of "stacked" trial and target population data}

\item{selection_method}{method to estimate the probability of trial participation.  Default is logistic regression ("lr").  Other methods supported are Random Forests ("rf") and Lasso ("lasso")}

\item{is_data_disjoint}{logical. If TRUE, then trial and population data are considered independent.  This affects calculation of the weights - see details for more information.}

\item{trim_pop}{logical. If TRUE, then population data are subset to exclude individuals with covariates outside bounds of trial covariates.}

\item{seed}{numeric. By default, the seed is set to 12222, otherwise can be specified (such as for simulation purposes).}
}
\description{
This function, given a stacked data frame containing both sample and population data, assesses the generalizability of the sample to the population on given covariates.
}
\details{
'assess_wrap()' is a wrapper for this function that allows assessment over levels of a grouping variable.
}
