% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/recruit.R
\name{.get.sample.size}
\alias{.get.sample.size}
\title{Internal function that asks user to specify sample size for guided version of \code{recruit()} and validates user-specified inputs}
\usage{
.get.sample.size(valid_inputs)
}
\arguments{
\item{valid_inputs}{vector, enumerates the possible number of units to recruit, based on the population size}
}
\value{
The function returns a list of formatted recruitment results for user interpretation: a table of results, a corresponding kable, and a the number of units selected from each stratum
}
\description{
Intended only to be called within \code{recruit()}, not as a standalone function
}
\keyword{internal}
