% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{.make.covariate.table}
\alias{.make.covariate.table}
\title{Create Covariate Balance Table}
\usage{
.make.covariate.table(
  data,
  sample_var,
  covariates,
  weighted_table = FALSE,
  estimation_method = "lr",
  is_data_disjoint = TRUE
)
}
\arguments{
\item{data}{data frame comprised of "stacked" sample and target population data}

\item{sample_var}{variable name denoting sample membership (1 = in sample, 0 = out of sample)}

\item{covariates}{vector of covariate names in data set that predict sample membership}

\item{weighted_table}{defaults to FALSE; whether weights are already included and do not need to be estimated}

\item{estimation_method}{method to estimate the probability of sample membership.  Default is logistic regression ("lr").  Other methods supported are Random Forests ("rf") and Lasso ("lasso")}

\item{is_data_disjoint}{defaults to TRUE. If TRUE, then sample and population data are considered independent.  This affects calculation of the weights}
}
\description{
This function is designed for use within \code{weighting()} and \code{assess()}.'
}
