% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stratify.R
\name{.stratify.unguided}
\alias{.stratify.unguided}
\title{Internal function that provides the unguided version of \code{stratify()}}
\usage{
.stratify.unguided(
  data,
  n_strata = NULL,
  variables = NULL,
  idvar = NULL,
  verbose = TRUE
)
}
\arguments{
\item{data}{data.frame object containing the population data to be stratified.}
}
\description{
Intended only to be called within \code{stratify}, not as a standalone function
}
\keyword{internal}
