% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stratify.R
\name{print.generalizer_stratify}
\alias{print.generalizer_stratify}
\title{Internal function that formats and prints a summary of stratification results.}
\usage{
\method{print}{generalizer_stratify}(x, ...)
}
\arguments{
\item{x}{generalizer_stratify, an object produced by the stratify() function including the dataset and parameters used to guide stratification as well as the stratification solution}
}
\value{
The function invisible returns the input generalizer_stratify object
}
\description{
Intended only to be called within \code{stratify_guided} and \code{stratify_unguided}, not as a standalone function
}
\keyword{internal}
