% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/assess.R
\name{print.summary.generalizeR_assess}
\alias{print.summary.generalizeR_assess}
\title{Print method for "summary.generalizeR_assess" class}
\usage{
\method{print}{summary.generalizeR_assess}(x, ...)
}
\arguments{
\item{x}{An object of class "summary.generalizeR_assess"}

\item{...}{Other arguments passed to or from other methods}
}
\value{
The generalizability index for the sample, the distribution of propensity scores, and a summary of the given covariates
}
\description{
Print method for "summary.generalizeR_assess" class
}
