% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stratify.R
\name{print.summary.generalizer_stratify}
\alias{print.summary.generalizer_stratify}
\title{Internal function that formats and prints a summary.generalizer_stratify object}
\usage{
\method{print}{summary.generalizer_stratify}(x, ...)
}
\arguments{
\item{x}{summary.generalizer_stratify, an object produced by the summary.generalizer_stratify() function}
}
\value{
The function invisible returns the inputted summary.generalizer_stratify object x
}
\description{
Intended only to be called within \code{stratify_guided} and \code{stratify_unguided}, not as a standalone function
}
\keyword{internal}
