% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stratify.R
\name{print.summary.generalizeR_stratify}
\alias{print.summary.generalizeR_stratify}
\title{Print method for "summary.generalizeR_stratify" class}
\usage{
\method{print}{summary.generalizeR_stratify}(x, ...)
}
\arguments{
\item{x}{An object of class "generalizeR_stratify"}

\item{...}{Other arguments passed to or from other methods}
}
\value{
A summary of the covariates used to stratify that compare them to the general population
}
\description{
Print method for "summary.generalizeR_stratify" class
}
