% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stratify.R
\name{summary.generalizer_stratify}
\alias{summary.generalizer_stratify}
\title{Internal function that assigns class generalizer_stratify to the inputted object}
\usage{
\method{summary}{generalizer_stratify}(object, ...)
}
\arguments{
\item{object}{generalizer_stratify, an object produced by the stratify() function including the dataset and parameters used to guide stratification as well as the stratification solution}
}
\value{
The function returns the input generalizer_stratify object as a summary.generalizer_stratify object
}
\description{
Intended only to be called within \code{stratify_guided} and \code{stratify_unguided}, not as a standalone function
}
\keyword{internal}
