% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stratify.R
\name{summary.generalizeR_stratify}
\alias{summary.generalizeR_stratify}
\title{Summary method for "generalizeR_stratify" class}
\usage{
\method{summary}{generalizeR_stratify}(x, ...)
}
\arguments{
\item{x}{An object of class "generalizeR_stratify"}

\item{...}{Other arguments passed to or from other methods}
}
\description{
Summary method for "generalizeR_stratify" class
}
