% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/weighting.R
\name{summary.generalizeR_weighting}
\alias{summary.generalizeR_weighting}
\title{Summary method for "generalizeR_weighting" class}
\usage{
\method{summary}{generalizeR_weighting}(object, ...)
}
\arguments{
\item{object}{An object of class "generalizeR_weighting"}

\item{...}{Other arguments passed to or from other methods}
}
\value{
A summary of the weighted and unweighted dataset for propensity scores, covariates and (optionally) TATE
}
\description{
Summary method for "generalizeR_weighting" class
}
