% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/weighting.R
\name{weighting}
\alias{weighting}
\title{Estimate Weights for Generalizing Average Treatment Effect}
\usage{
weighting(
  data,
  sample_indicator,
  treatment_indicator = NULL,
  outcome = NULL,
  covariates,
  estimation_method = "lr",
  disjoint_data = TRUE
)
}
\arguments{
\item{data}{data frame comprised of "stacked" sample and target population data}

\item{sample_indicator}{variable name denoting binary sample membership (1 = in sample, 0 = out of sample)}

\item{treatment_indicator}{variable name denoting binary treatment assignment (ok if only available in sample, not population)}

\item{outcome}{variable name denoting outcome}

\item{covariates}{vector of covariate names in data set that predict sample membership}

\item{estimation_method}{method to estimate the probability of sample membership. Default is logistic regression ("lr").Other methods supported are Random Forests ("rf") and Lasso ("lasso")}

\item{disjoint_data}{logical. If TRUE, then sample and population data are considered disjoint. This affects calculation of the weights - see details for more information.}
}
\description{
This function is designed for use within 'covariate_table()' and 'assess()'.
}
